python-analytics/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
apscheduler==3.10.4
python-dotenv==1.0.1
